requirements.txt
discord.py>=2.6.0
python-dotenv>=1.2.0
aiohttp>=3.13.0
asyncpg>=0.31.0
PyNaCl>=1.6.2
